+++ solaris_oci/cli/image/__init__.py
@@ -13,6 +13,3 @@
 # limitations under the License.
 
 from .image import Image
-from .hydrate import Hydrate
-from .inspect import Inspect
-from .list import List