+++ solaris_oci/cli/image/inspect.py
@@ -12,20 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import subprocess
-import argparse
-import pathlib
+
 import json
-from dateutil import parser
-import humanize
-from datetime import *
-
-from opencontainers.digest import NewDigestFromEncoded, Parse
-from opencontainers.image.v1 import Index
-
-from solaris_oci.util.print import print_table
+import argparse
 from solaris_oci.oci.image import Distribution
-from solaris_oci.oci import config
 
 class Inspect:
     @staticmethod
@@ -36,15 +26,42 @@ class Inspect:
             description='Display detailed information on one or more images',
             help='Display detailed information on one or more images')
         parser.add_argument('image',
-            nargs='+',
+            nargs='+', 
             metavar='IMAGE',
             help='Name of the image to inspect')
  
     def __init__(self, options):
-        self.options = options
-        oci_path = pathlib.Path(options.root)
+        images = self.images(options.image)
+        print(json.dumps(images, indent=4, default=str))
 
+    def images(self, filter):
+        repositories = {}
+        for reference in filter:
+            records = reference.split(':')
+            name = records[0]
+            tag = None
+            if len(records) == 2 and len(records[1]) != 0:
+                tag = records[1]
+            tags = repositories.get(name, [])
+            if tag is None:
+                value = None
+            else:
+                if tags is None:
+                    value = None
+                else:
+                    tags.append(tag)
+                    value = tags
+            repositories[name] = value
+    
         distribution = Distribution()
         distribution.load()
-        images = distribution.images(options.image)
-        print(json.dumps(images, indent=4, default=str))
+        images = []
+        for repository in distribution.repositories:
+            if repository.name in repositories.keys():
+                tags = repositories[repository.name]
+                for tag in repository.images:
+                    if tags is None or tag.tag in tags:
+                        for manifest in tag.manifests:
+                            image = manifest['config']
+                            images.append(image.to_dict(use_real_name=True))
+        return images