+++ solaris_oci/oci/image/distribution.py
@@ -12,46 +12,25 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import json 
+from opencontainers.distribution.v1 import RepositoryList
 from .. import config as osi_config
-from .repository import Repository
-from .descriptor import Descriptor
-import json
+from . import Repository
 
 class Distribution():
-    def __init__(self, descriptor_json=None):
+    def __init__(self):
+        self.repository_list = None
         self.repositories = None
-    
+
     def load(self):
-        with osi_config.distribution_json_path.open() as distribution_file:
-            # load distribution.json
-            distribution_json = json.load(distribution_file)
-            self.schemaVersion = distribution_json['schemaVersion']
+        distribution_path = osi_config.distribution_path
+        distribution_json_path = osi_config.distribution_json_path
+        with distribution_json_path.open() as distribution_json_file:
+            distribution_json = json.load(distribution_json_file)
+            self.repository_list = RepositoryList()
+            self.repository_list.load(distribution_json)
             self.repositories = []
-            for repository_descriptor_json in distribution_json['repositories']:
-                self.repositories.append(Repository(repository_descriptor_json))
-
-    def images(self, filter=None):
-        images = []
-        for repository in self.repositories:
-            for index in repository.indexes:
-                for manifest in index.manifests:
-                    image = manifest.image.data.copy()
-                    image['registry'] = repository.registry
-                    image['repository'] = repository.name
-                    image['id'] = repository.descriptor.digest
-                    image['digest'] = index.descriptor.digest
-                    image['tag'] = index.name
-                    image['size'] = manifest.size
-                    images.append(image)
-        
-        if filter is not None:
-            names = []
-            for reference in filter:
-                records = reference.split(':')
-                name = records[0]
-                if len(records) == 2:
-                    raise Exception('NYI')
-                names.append(name)
-            images = [i for i in images if i['repository'] in names]
-
-        return images
+            for repository_name in self.repository_list.get('Repositories'):
+                repository = Repository(repository_name)
+                repository.load(distribution_path)
+                self.repositories.append(repository)